Removed unnecessary comments

+++ alcohol.py
@@ -1,15 +1,12 @@
+
 ## Program Name: alchohol.py
 ## Programmer: Patrick Will
 ## Date: 6/16/2018
 ## Purpose: A program to find the alcohol content in ounces of an alcholic drink
 
 #variables
-#abv = 0
-#volume = 0
-
-#user input
 abv = float(input('What is the ABV in percent?: '))
-volume = int(input('What is the volume in ounces?: '))
+volume = float(input('What is the volume in ounces?: '))
 
 #functions
 def CalculateTotalAlcohol(v,a):
